% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_dist.R
\name{mid_dist}
\alias{mid_dist}
\title{Mid-distance function}
\usage{
mid_dist(d1, d2, theta, Dmax)
}
\arguments{
\item{d1}{Distance of one bridge from apex.}

\item{d2}{Distance of other bridge from apex.}

\item{theta}{Angle of apex (in radians)}

\item{Dmax}{Maximum possible distance of a bridge from apex.}
}
\value{
mid-distance
}
\description{
This function models the distance between two
             bridges along a common arm.
}
\author{
Jason Graham, \email{jason.graham@scranton.edu}

 @export
}
