% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_density_general.R
\name{skew_density_general}
\alias{skew_density_general}
\title{Density Function for Asymmetric Apparatus}
\usage{
skew_density_general(d, angle1, angle2, ant.dens = 2.2, LT = 100,
  LS0 = 22, LL0 = 44, wA = 3.3, ln = 0.691, wn = 0.107,
  alpha = 17.02)
}
\arguments{
\item{d}{vector of distances values}

\item{angle1}{Angle left arm makes with vertical (in degrees)}

\item{angle2}{Angle right arm makes with vertical (in degrees)}

\item{ant.dens}{Density of ants, default value =  2.2}

\item{LT}{Overall trail length, default value = 100 (cm)}

\item{LS0}{Length of left arm, default value = 22 (cm)}

\item{LL0}{Length of right arm, default value = 44 (cm)}

\item{wA}{Width of apparatus, default value = 3.3 (cm)}

\item{ln}{Length of avg ant default value = 0.691 (cm)}

\item{wn}{Width of avg ant default value = 0.107 (cm)}

\item{alpha}{Free fitting parameter value from Reid et al, default value 17.02}
}
\value{
Density \code{rho}
}
\description{
This function models the density as a function of
             distance for the asymmetric appratus.
}
\author{
Jason Graham, \email{jason.graham@scranton.edu}

 @export
}
