% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_density_general.R
\name{two_density_general}
\alias{two_density_general}
\title{Density Function for Two-Apex Apparatus}
\usage{
two_density_general(d, angle, ant.dens = 2.2, LT = 100, L0 = 22,
  wA = 3.3, ln = 0.691, wn = 0.107, alpha = 17.02)
}
\arguments{
\item{d}{A numeric dector of distances values in cm.}

\item{angle}{A vector of angles of apex in deg.}

\item{ant.dens}{A scalar for the density of ants in ants/cm (default: 2.2).}

\item{LT}{A scalar for the overall trail length in cm (default: 100).}

\item{L0}{A scalar for the length of the apparatus arm in cm (default: 22).}

\item{wA}{A scalar for the width of the arms of the apparatus in cm (default: 3.3).}

\item{ln}{A scalar for the average length of an ant in cm (default:0.691).}

\item{wn}{A scalar for the average width of an ant in cm (default: 0.107).}

\item{alpha}{A scalar for the free fitting parameter value from Reid et al.
(default: 17.02).}
}
\value{
A vector of the of the same length as \code{d, angle}.
}
\description{
This function computes the density as a function of distance for
 the two-apex appratus.
}
\author{
Jason Graham, \email{jason.graham@scranton.edu}

Simon Garnier, \email{garnier@njit.edu}
}
