% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_dist.R
\name{mid_dist}
\alias{mid_dist}
\title{Mid-Distance Function}
\usage{
mid_dist(d1, d2, theta, Dmax)
}
\arguments{
\item{d1}{A numeric vector of distances of first bridge from apex in cm.}

\item{d2}{A numeric vector of distances of second bridge from apex in cm.}

\item{theta}{A numeric vector of angles of apex in rad.}

\item{Dmax}{A numeric vector of the maximum possible distances of a bridge
from apex in cm.}
}
\value{
A vector of the same length as \code{d1, d2, theta}.
}
\description{
This function models the distance between two bridges along a
 common arm.
}
\author{
Jason Graham, \email{jason.graham@scranton.edu}
}
