% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_density_general.R
\name{skew_density_general}
\alias{skew_density_general}
\title{Density Function for Asymmetric Apparatus}
\usage{
skew_density_general(d, angle1, angle2, ant.dens = 2.2, LT = 100,
  LS0 = 22, LL0 = 44, wA = 3.3, ln = 0.691, wn = 0.107,
  alpha = 17.02)
}
\arguments{
\item{d}{A length 2 numeric vector of distances values in cm.}

\item{angle1}{A scalar for the angle in deg that the left arm makes with
vertical.}

\item{angle2}{A scalar for the angle in deg that the right arm makes with
vertical.}

\item{ant.dens}{A scalar for the density of ants in ants/cm (default: 2.2).}

\item{LT}{A scalar for the overall trail length in cm (default: 100).}

\item{LS0}{A scalar for the length of the left arm in cm (default: 22).}

\item{LL0}{A scalar for the length of the right arm in cm (default: 44).}

\item{wA}{A scalar for the width of the arms of the apparatus in cm (default: 3.3).}

\item{ln}{A scalar for the average length of an ant in cm (default:0.691).}

\item{wn}{A scalar for the average width of an ant in cm (default: 0.107).}

\item{alpha}{A scalar for the free fitting parameter value from Reid et al.
(default: 17.02).}
}
\value{
A vector of the of the same length as \code{d}.
}
\description{
This function computes the density as a function of distance for
 the asymmetric appratus.
}
\author{
Jason Graham, \email{jason.graham@scranton.edu}

Simon Garnier, \email{garnier@njit.edu}
}
