% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_density_general.R
\name{three_density_general}
\alias{three_density_general}
\title{Density Function for Three-apex Apparatus}
\usage{
three_density_general(d, angle, ant.dens = 2.2, LT = 100, L0 = 22,
  ln = 0.691, wn = 0.107, wA = 3.3, alpha = 17.02)
}
\arguments{
\item{d}{vector of distances values}

\item{angle}{Angle of apex (in degrees)}

\item{ant.dens}{Density of ants, default value =  2.2}

\item{LT}{Overall trail length, default value = 100 (cm)}

\item{L0}{Length of one arm, default value = 22 (cm)}

\item{ln}{Length of avg ant default value = 0.691 (cm)}

\item{wn}{Width of avg ant default value = 0.107 (cm)}

\item{wA}{Width of apparatus, default value = 3.3 (cm)}

\item{alpha}{Free fitting parameter value from Reid et al, default value 17.02}
}
\value{
Density \code{rho}
}
\description{
This function models the density as a function of
             distance for the two-apex appratus.
}
\author{
Jason Graham, \email{jason.graham@scranton.edu}

 @export
}
